Refactor: let those page load immediately
