+++ app/mypage/change-pw/loading.tsx
@@ -1,4 +1,5 @@
-// TODO
+import ChangePWForm from "./form";
+
 export default function ChangePWLoading() {
-  return <div />;
+  return <ChangePWForm input={{old_password : '', new_password : '', new_password_check : ''}} submitting={false} />;
 }
