+++ app/mypage/change-pw/page.tsx
@@ -1,9 +1,51 @@
-import ChangePWLoading from './loading';
+'use client';
 
-import dynamic from 'next/dynamic';
+import { apiProfileUpdatePW } from '@/lib/api/mypage';
+import { useState } from 'react';
+import { useSessionId } from '@/context/SessionIdContext';
+import { retryWithJWTRefresh } from '@/lib/api/authHelper';
 
-const ChangePW = dynamic(() => import('./client'), { ssr: false, loading: ChangePWLoading });
+import { useRouter } from 'next/navigation';
 
-export default function ChangePWPage() {
-  return <ChangePW />;
+import ChangePwForm from './form';
+import { handleInputBuilder } from '@/lib/form/handler';
+
+export default function ChangePW() {
+  const [input, setInput] = useState<UpdatePWExtended>({ old_password: '', new_password: '', new_password_check: '' });
+
+  const [busy, setBusy] = useState<boolean>(false);
+
+  const sessionIdState = useSessionId();
+  const router = useRouter();
+
+  function handleSubmit(e: React.FormEvent) {
+    e.preventDefault();
+    setBusy(true);
+    if (input.new_password !== input.new_password_check) {
+      setBusy(false);
+      return alert('비밀번호와 비밀번호 확인이 일치하지 않습니다.');
+    }
+
+    retryWithJWTRefresh(apiProfileUpdatePW, sessionIdState)(
+      { old_password: input.old_password, new_password: input.new_password },
+      {},
+    ).then((s) => {
+      if (s.status === 'SUCCESS') {
+        alert(`비밀번호 변경에 성공했습니다.`);
+        router.push('/mypage?pwchanged');
+      } else {
+        alert(`변경에 실패했습니다: ${s.message}`);
+      }
+    });
+    setBusy(false);
+  }
+
+  return (
+    <ChangePwForm
+      handleSubmit={handleSubmit}
+      handleInput={handleInputBuilder(input, setInput)}
+      input={input}
+      submitting={busy}
+    />
+  );
 }
