+++ app/mypage/delete-account/loading.tsx
@@ -1,0 +1,7 @@
+import DeleteAccountForm from "./form";
+
+export default function DeleteAccountLoading() {
+  return (
+    <DeleteAccountForm input={{password : '', checked : false}} submitting={false} />
+  );
+}
